Rename the svg folder to renderers folder
