+++ src/renderers/diag-to-svg.ts
