+++ src/renderers/svg-renderer.ts
