+++ src/renderers/text-renderer.ts
